Fix the four defects adversarial review found in the gamedata slice
1. Float ABI (would have made the live gate silently no-op). The invoke thunk
   typed its 8 xmm slots `double` while the vocabulary's `float` is 32-bit, so
   SysV handed the callee a double bit pattern: (double)10.0 has a zero low
   word, and a callee doing movss read 0.0f. No crash, no diagnostic — the exact
   silent-misbehaviour class this slice exists to prevent. This came from the
   PLAN, which specified `double`. Slots are now `float` with an explicit
   f64->f32 narrow, and scripts/check-invoke-abi.sh gates it two ways: it greps
   the real prototypes AND compiles a separate-TU -O0 round-trip proving a float
   argument survives. Verified the gate fails if the types regress.

2. Call-name injection defeated the type gate entirely. Names are interpolated
   verbatim into the generated .d.ts, and nothing validated them, so a key of
   `y: (...a: any[]) => any; [k: string]: any; z` injected an index signature —
   after which every Engine.call("anything") typechecked while degrading to null
   at runtime. Names must now be plain identifiers; `constructor`/`prototype`/
   `__proto__` are reserved.

3. The `vector` arg type was inverted. It generated a tuple, but core reads
   .x/.y/.z properties (defaulting to 0.0) — so passing a real Vector FAILED
   typecheck while `[1,2,3]` typechecked and silently sent (0,0,0). Now the
   structural {x,y,z} shape, which Vector and QAngle satisfy and arrays don't.
   It had zero test coverage, which is why it was wrong; now covered.

4. make ci-js was RED. check-examples-coverage.sh requires every shipped SDK
   subpath to have a consumer, and ./unsafe had none. Adds a cookbook recipe
   that also demonstrates the default-deny path, using the RE note's VERIFIED
   Ignite descriptor (CBaseModelEntity, not CBaseEntity — index 264 is past the
   end of CBaseEntity's 237-slot vtable).

Also: corrected the RE note's overrider census (15 across two implementations,
not 11 across one — independently confirmed via link-time vtable reference
counts at a 240/11/4 ratio; a negative fixture built from CPhysicsProp resolves
to 0x19836c0, not the documented 0x1628d30), gave the build test's bare
assert.rejects a TS2554 matcher so it can tell a working gate from a dead one,
and added the missing changeset for the new public subpath.

Gates: SDK 204 tests / 203 pass (the 1 failure is pre-existing and
environmental — packages/eslint-plugin has no dist/); check-plugins-typecheck,
check-examples-coverage, check-invoke-abi all green; make shim links clean.
Proved the augmentation is live, not silently `never`: an undeclared call name
is TS2345 and a wrong arity is TS2554.

+++ shim/src/engine_calls.cpp
@@ -168,10 +168,18 @@ int Fail(char* out, int cap, const char* reason) {
 
 // The two max-arity prototypes (see THE INVOKE TECHNIQUE above). All 5 GP + 8 xmm slots are always
 // passed; the callee reads only what its real prototype declares.
+//
+// The xmm slots MUST be `float`, not `double`. The arg vocabulary's `float` is 32-bit, and SysV
+// passes a 32-bit float as a float in the LOW 32 bits of the xmm register. Declaring these `double`
+// puts a 64-bit double bit pattern in the register, so a callee doing `movss`/`movd` reads the
+// double's low word: `(double)10.0` is 0x4024000000000000, whose low 32 bits are ZERO, so the callee
+// sees 0.0f. That is a SILENT wrong value — no crash, no diagnostic — and it is exactly the
+// misbehaviour class this slice exists to prevent. Verified by compiled repro; guarded by the
+// float-round-trip test in core/src/gamedata_calls.rs.
 using FnU64 = uint64_t (*)(void*, uint64_t, uint64_t, uint64_t, uint64_t, uint64_t,
-                           double, double, double, double, double, double, double, double);
+                           float, float, float, float, float, float, float, float);
 using FnF32 = float    (*)(void*, uint64_t, uint64_t, uint64_t, uint64_t, uint64_t,
-                           double, double, double, double, double, double, double, double);
+                           float, float, float, float, float, float, float, float);
 
 }  // namespace
 
@@ -276,7 +284,9 @@ int S2_EngineCallInvoke(int callId, int entIndex, int entSerial, int subObjOff,
     }
 
     uint64_t g[kMaxGpArgs] = { 0, 0, 0, 0, 0 };
-    double   f[kMaxFpArgs] = { 0, 0, 0, 0, 0, 0, 0, 0 };
+    // `float`, not `double` — see the FnU64/FnF32 note above. The core hands us f64 across the ABI
+    // (JS numbers are doubles); the narrowing to the engine's 32-bit float happens HERE, once.
+    float    f[kMaxFpArgs] = { 0, 0, 0, 0, 0, 0, 0, 0 };
 
     for (int i = 0; i < gpCount; i++) {
         switch (gpKind[i]) {
@@ -311,7 +321,7 @@ int S2_EngineCallInvoke(int callId, int entIndex, int entSerial, int subObjOff,
                 return 0;                            // unknown arg kind — degrade, never guess
         }
     }
-    for (int i = 0; i < fpCount; i++) f[i] = fp[i];
+    for (int i = 0; i < fpCount; i++) f[i] = static_cast<float>(fp[i]);   // explicit f64 -> f32 narrow
 
     if (retKind == kRetFloat) {
         float r = reinterpret_cast<FnF32>(fn)(thisPtr, g[0], g[1], g[2], g[3], g[4],
